Add velocity smoothing utilities and evaluation updates
