+++ MATLAB/evaluate_filter_results.m
@@ -1,0 +1,17 @@
+function evaluate_filter_results(~)
+%EVALUATE_FILTER_RESULTS  Stub for Task 7 evaluation in MATLAB.
+%
+% Usage:
+%   evaluate_filter_results(npz_file, output_dir, tag)
+%
+% This stub mirrors ``evaluate_filter_results.py``. It should load the
+% Kalman filter output from a ``*_kf_output.npz`` file, compute residual
+% statistics and plot attitude angles and error norms.
+%
+% TODO: implement MATLAB version matching the Python functionality.
+%
+% Inputs are intentionally unused until implemented.
+
+warning('evaluate_filter_results:notImplemented', ...
+    'MATLAB version not yet implemented. See evaluate_filter_results.py.');
+end
